requirements.txt
# Reticulum Network Stack
rns==1.1.4
lxmf==0.9.4

# UI Framework
kivy==2.3.1
kivymd==1.2.0
mapview==1.0.6

# Encryption
sqlcipher3==0.6.2
pynacl==1.6.2
argon2-cffi==25.1.0

# Configuration
pyyaml>=6.0

# Development
pytest>=8.0.0
buildozer>=1.5.0
pyinstaller>=6.0.0
